+++ gm-domain/src/main/java/com/gm/modules/basicconfig/entity/GmTeamConfigEntity.java
@@ -69,7 +69,7 @@ public class GmTeamConfigEntity implements Serializable {
 	/**
 	 * 队伍战力
 	 */
-	private Long teamPower;
+	private Double teamPower;
 	/**
 	 * 队伍矿工数
 	 */
